add new module (photo-ui)
